Añadido proyecto base de Angular
Implementadas las pantallas de calcular huella  para transporte privado
y para vuelos

+++ CarbonCalcutaorBackEnd/src/main/java/net/atos/zerokhoi/controller/CarbonInterfaceController.java
@@ -28,7 +28,7 @@ public class CarbonInterfaceController {
 	}
 
 	@PostMapping(value = "/huella")
-	private Integer obtenerHuella(@RequestBody DataElectricidad data) {
+	private Double obtenerHuella(@RequestBody DataElectricidad data) {
 		
 		String uri = "https://www.carboninterface.com/api/v1/estimates";		
 		UriComponentsBuilder builderUrl = UriComponentsBuilder.fromHttpUrl(uri);
